+++ Assets/RCore.LXLite/Editor/LXLiteWindow.cs
@@ -94,7 +94,7 @@ namespace RCore.LXLite.Editor
 		[MenuItem("Window/" + MENU)]
 #else
 		public const int GROUP_14 = 140;
-		[MenuItem("RCore/Tools/" + MENU, priority = GROUP_14)]
+		[MenuItem("RCore/" + MENU, priority = 25)]
 #endif
 		public static void ShowWindow()
 		{